style: 주석 정리, 클래스 명 변경
