+++ src/main/java/com/zerobase/funding/api/config/WebConfig.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-//@EnableWebMvc
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
